refactor(brutus): extract config types to config.go
