+++ cmd/brutus/config.go
@@ -1,0 +1,58 @@
+// Copyright 2026 Praetorian Security, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package main
+
+import (
+	"time"
+
+	"github.com/praetorian-inc/brutus/pkg/brutus"
+)
+
+// baseConfigOptions holds common configuration shared across targets
+type baseConfigOptions struct {
+	usernames         []string
+	passwords         []string
+	keys              [][]byte
+	threads           int
+	timeout           time.Duration
+	stopOnSuccess     bool
+	snmpTier          string
+	llmConfig         *brutus.LLMConfig
+	browserTimeout    time.Duration
+	browserTabs       int
+	browserVisible    bool
+	useHTTPS          bool
+	useColor          bool
+	quiet             bool
+	verbose           bool
+	useBadkeys        bool
+	protocolOverride  string              // Override fingerprintx-detected protocol
+	aiMode            bool                // Enable AI-powered credential detection for HTTP
+	aiVerify          bool                // Use Claude Vision to verify login success
+	aiResearchedCreds []brutus.Credential // Credentials researched by LLM for current target
+	tlsMode           string              // TLS verification mode: "disable", "verify", "skip-verify"
+	rateLimit         float64             // Max requests per second (0 = unlimited)
+	jitter            time.Duration       // Random delay variance for rate limiting
+	maxAttempts       int
+	sprayMode         bool
+}
+
+// determineTLSMode returns the appropriate TLS mode based on the verify-tls flag
+func determineTLSMode(verifyTLS bool) string {
+	if verifyTLS {
+		return "verify"
+	}
+	return "disable"
+}
