+++ cmd/brutus/main.go
@@ -329,43 +329,6 @@ func main() {
 	}
 }
 
-// baseConfigOptions holds common configuration shared across targets
-type baseConfigOptions struct {
-	usernames         []string
-	passwords         []string
-	keys              [][]byte
-	threads           int
-	timeout           time.Duration
-	stopOnSuccess     bool
-	snmpTier          string
-	llmConfig         *brutus.LLMConfig
-	browserTimeout    time.Duration
-	browserTabs       int
-	browserVisible    bool
-	useHTTPS          bool
-	useColor          bool
-	quiet             bool
-	verbose           bool
-	useBadkeys        bool
-	protocolOverride  string              // Override fingerprintx-detected protocol
-	aiMode            bool                // Enable AI-powered credential detection for HTTP
-	aiVerify          bool                // Use Claude Vision to verify login success
-	aiResearchedCreds []brutus.Credential // Credentials researched by LLM for current target
-	tlsMode           string              // TLS verification mode: "disable", "verify", "skip-verify"
-	rateLimit         float64             // Max requests per second (0 = unlimited)
-	jitter            time.Duration       // Random delay variance for rate limiting
-	maxAttempts       int
-	sprayMode         bool
-}
-
-// determineTLSMode returns the appropriate TLS mode based on the verify-tls flag
-func determineTLSMode(verifyTLS bool) string {
-	if verifyTLS {
-		return "verify"
-	}
-	return "disable"
-}
-
 // hasStdinData checks if stdin has data available (i.e., is being piped to)
 func hasStdinData() bool {
 	stat, err := os.Stdin.Stat()
